feat: improve publish settings
